adds favorites-empty component, change main page class names
