+++ src/pages/main-page/main-page.tsx
@@ -15,7 +15,7 @@ function MainPage() {
   const activeOffers = offers.filter((offer) => offer.city.name === activeCity);
   const [selectedCard, setSelectedCard] = useState('');
 
-  const mainIndexClasses = cn('page__main--index', {
+  const mainClasses = cn('page__main--index', {
     'page__main--index-empty': activeOffers.length === 0
   });
   const citiesPlacesClasses = cn('cities__places-container container', {
@@ -26,7 +26,7 @@ function MainPage() {
     <Layout
       title='Main'
       pageClass='page--gray page--main'
-      mainClass={mainIndexClasses}
+      mainClass={mainClasses}
     >
       <>
         <h1 className="visually-hidden">Cities</h1>
@@ -44,7 +44,7 @@ function MainPage() {
               : <CitiesNoPlaces city={activeCity}/>}
 
             <div className="cities__right-section">
-              {activeOffers.length &&
+              {activeOffers.length > 0 &&
                 <Map
                   activeCity={activeOffers[0].city}
                   points={getMapPoints(activeOffers)}
